+++ Tests/ASCollectionViewTests.mm
@@ -1050,8 +1050,17 @@
   // Trigger the initial reload to start
   [window layoutIfNeeded];
 
+  // Test the APIs that monitor ASCollectionNode update handling
+  XCTAssertTrue(cn.isProcessingUpdates, @"ASCollectionNode should still be processing updates after initial layoutIfNeeded call (reloadData)");
+  [cn onDidFinishProcessingUpdates:^{
+    XCTAssertTrue(!cn.isProcessingUpdates, @"ASCollectionNode should no longer be processing updates inside -onDidFinishProcessingUpdates: block");
+  }];
+
   // Wait for ASDK reload to finish
   [cn waitUntilAllUpdatesAreCommitted];
+
+  XCTAssertTrue(!cn.isProcessingUpdates, @"ASCollectionNode should no longer be processing updates after -wait call");
+
   // Force UIKit to read updated data & range controller to update and account for it
   [cn.view layoutIfNeeded];